chore: E2E tests refactor
