+++ edc-tests/edc-controlplane/fixtures/src/testFixtures/java/org/eclipse/tractusx/edc/tests/ParticipantEdrApi.java
@@ -25,6 +25,7 @@ import jakarta.json.Json;
 import jakarta.json.JsonArray;
 import jakarta.json.JsonObject;
 import org.eclipse.edc.test.system.utils.Participant;
+import org.eclipse.tractusx.edc.tests.participant.TransferParticipant;
 
 import static io.restassured.http.ContentType.JSON;
 import static jakarta.json.Json.createObjectBuilder;
@@ -109,7 +110,7 @@ public class ParticipantEdrApi {
      * @param callbacks The callbacks
      * @return The contract negotiation id
      */
-    public String negotiateEdr(TxParticipant other, String assetId, JsonArray callbacks) {
+    public String negotiateEdr(TransferParticipant other, String assetId, JsonArray callbacks) {
         var dataset = participant.getDatasetForAsset(other, assetId);
         assertThat(dataset).withFailMessage("Catalog received from " + other.getName() + " was empty!").isNotEmpty();
