+++ edc-tests/edc-controlplane/fixtures/src/testFixtures/java/org/eclipse/tractusx/edc/tests/participant/TransferParticipant.java
@@ -1,0 +1,43 @@
+/********************************************************************************
+ * Copyright (c) 2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License, Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ********************************************************************************/
+
+package org.eclipse.tractusx.edc.tests.participant;
+
+/**
+ * Extension of {@link TractusxParticipantBase} with Transfer specific configuration
+ */
+public class TransferParticipant extends TractusxParticipantBase {
+
+    public static class Builder extends TractusxParticipantBase.Builder<TransferParticipant, Builder> {
+
+        protected Builder() {
+            super(new TransferParticipant());
+        }
+
+        public static Builder newInstance() {
+            return new Builder();
+        }
+
+        @Override
+        public TransferParticipant build() {
+            super.build();
+            return participant;
+        }
+    }
+}
